# Speed up is_ascii_hex_digit