@anonrig 2023-08-22
Can you run and share the results of the benchmarks?