@lemire 2023-08-22
I am using GCC 11 on an Ice Lake-based server.


## benchdata



Main branch...

```
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
BasicBench_AdaURL_href              37300975 ns     37258253 ns           19 GHz=3.18743 cycle/byte=13.644 cycles/url=1.1851k instructions/byte=37.2856 instructions/cycle=2.73275 instructions/ns=8.71046 instructions/url=3.2386k ns/url=371.806 speed=233.186M/s time/byte=4.28843ns time/url=372.489ns url/s=2.68464M/s
BasicBench_AdaURL_aggregator_href   23117267 ns     23089445 ns           30 GHz=3.1885 cycle/byte=8.47386 cycles/url=736.033 instructions/byte=25.2896 instructions/cycle=2.98443 instructions/ns=9.51583 instructions/url=2.19663k ns/url=230.84 speed=376.28M/s time/byte=2.6576ns time/url=230.837ns url/s=4.33207M/s
```

This PR:

```
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
BasicBench_AdaURL_href              37287069 ns     37245642 ns           19 GHz=3.18801 cycle/byte=13.6441 cycles/url=1.18512k instructions/byte=37.2856 instructions/cycle=2.73272 instructions/ns=8.71193 instructions/url=3.23859k ns/url=371.742 speed=233.265M/s time/byte=4.28698ns time/url=372.363ns url/s=2.68555M/s
BasicBench_AdaURL_aggregator_href   23225765 ns     23195167 ns           30 GHz=3.1879 cycle/byte=8.50118 cycles/url=738.406 instructions/byte=25.2641 instructions/cycle=2.97184 instructions/ns=9.47392 instructions/url=2.19442k ns/url=231.628 speed=374.565M/s time/byte=2.66977ns time/url=231.894ns url/s=4.31232M/s
```

## bench_search_params

Main branch...

```

-----------------------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------
url_search_params_AdaURL      55819 ns        55726 ns        10737 GHz=3.19652 cycle/byte=11.596 cycles/url=32.4862k instructions/byte=40.3667 instructions/cycle=3.48108 instructions/ns=11.1273 instructions/url=113.087k ns/url=10.163k speed=201.092M/s time/byte=4.97284ns time/url=13.9314us url/s=71.7803k/s
```


This PR:
```
-----------------------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------
url_search_params_AdaURL      60966 ns        60869 ns        11080 GHz=3.19684 cycle/byte=10.7248 cycles/url=30.0455k instructions/byte=40.6287 instructions/cycle=3.7883 instructions/ns=12.1106 instructions/url=113.821k ns/url=9.3985k speed=184.099M/s time/byte=5.43185ns time/url=15.2173us url/s=65.7146k/s
```

## Conclusion

It does not appear that this PR improves the performance, at least according to these results.

However, the PR does suggest that we can do better.