@lemire 2023-08-22
@ttsugriy  Can you try...

```
cmake -B build -D ADA_BENCHMARKS=ON
cmake --build build
./build/benchmarks/bench_search_params
```