@ttsugriy 2023-08-22
what's even more interesting is that GCC 13 generates an even better assembly for the original version (https://compiler-explorer.com/z/1j9oMTrbc)
```
is_ascii_hex_digit(char):
        sub     edi, 48
        xor     eax, eax
        cmp     dil, 54
        ja      .L1
        movabs  rax, 35465847073801215
        bt      rax, rdi
        setc    al
.L1:
        ret
```