@anonrig 2023-08-22
> @lemire the reason you're not seeing any improvement is because you're using gcc 11 that produces a very different assembly (https://compiler-explorer.com/z/16eEbPezb) from the one I've shared above using clang 16.

@ttsugriy Can you share your benchmark results with clang 16?