@ttsugriy 2023-08-27
> > @lemire the reason you're not seeing any improvement is because you're using gcc 11 that produces a very different assembly (https://compiler-explorer.com/z/16eEbPezb) from the one I've shared above using clang 16.
> 
> @ttsugriy Can you share your benchmark results with clang 16?

@anonrig I have only clang 15 on my machine and the results are not very promising :)
main:
```
Running ./build/benchmarks/bench_search_params
Run on (8 X 24.0427 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB (x8)
  L1 Instruction 128 KiB (x8)
  L2 Unified 4096 KiB (x2)
Load Average: 3.73, 3.89, 3.53
bytes/URL: 2801.500000
input bytes: 11206
number of URLs: 4
performance counters: No privileged access (sudo may help).
-----------------------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------
url_search_params_AdaURL      60745 ns        60725 ns         9986 speed=184.538M/s time/byte=5.41895ns time/url=15.1812us url/s=65.871k/s
```
this PR
```
Running ./build/benchmarks/bench_search_params
Run on (8 X 24.1212 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB (x8)
  L1 Instruction 128 KiB (x8)
  L2 Unified 4096 KiB (x2)
Load Average: 3.18, 3.66, 3.47
bytes/URL: 2801.500000
input bytes: 11206
number of URLs: 4
performance counters: No privileged access (sudo may help).
-----------------------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------
url_search_params_AdaURL      63980 ns        63980 ns         9582 speed=175.147M/s time/byte=5.70948ns time/url=15.9951us url/s=62.5192k/s
```
clang info:
```
% clang --version
Apple clang version 15.0.0 (clang-1500.0.40.1)
```
As such this PR should probably be closed and it's probably best to try to write some sort of lookup table implementation so that clang generates the same super efficient assembly as GCC.