@ttsugriy 2023-08-27
The closest I could get with clang is by using
```
bool is_ascii_hex_digit(const char c) noexcept { 
    const unsigned ch = static_cast<unsigned>(c) - '0';
    if (ch > '6') return false;
    constexpr static bool table[] = {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
                                         0, 0, 0, 0, 0, 0, 0, 1, 1, 
                                         1, 1, 1, 1, 0, 0, 0, 0, 
                                          0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
                                          0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
                                          0, 0, 1, 1, 1, 1, 1, 1};
    return table[ch];
}
```
that generates
```
is_ascii_hex_digit(char):                # @is_ascii_hex_digit(char)
        add     edi, -48
        cmp     edi, 55
        setb    cl
        movabs  rax, 35465847073801215
        bt      rax, rdi
        setb    al
        and     al, cl
        ret
```
https://compiler-explorer.com/z/b93bYhn4j

but it's still much slower on the benchmark running on M1 ARM.